Rename contracts and add new test
